extracted css styles for components into individual files
