+++ frontend/src/features/canvas/components/ClearCanvasButton.module.css
@@ -1,0 +1,15 @@
+.clearCanvasButton {
+  position: absolute;
+  top: 35px;
+  right: 35px;
+  z-index: 9999;
+  padding: 10px 10px;
+  background-color: red;
+  color: white;
+  border: none;
+  border-radius: 8px;
+  cursor: pointer;
+  font-weight: 500;
+  font-size: 0.95rem;
+  width: max-content;
+}
