+++ frontend/src/features/canvas/components/confirmationModal.tsx
