+++ frontend/src/features/editors/boxEditor/BoxEditorStyles.module.css
@@ -132,44 +132,6 @@
   background-color: #059669;
 }
 
-/* Remove Button */
-.removeButton {
-  background-color: #ef4444;
-  color: white;
-}
-
-.removeButton:hover {
-  background-color: #dc2626;
-}
-
-/* ===== Invalidate Button ===== */
-.invalidateButton {
-  background-color: #3b82f6;
-  color: white;
-  padding: 10px 26px;
-  font-size: 0.95rem;
-  border-radius: 8px;
-  border: none;
-  cursor: pointer;
-  height: 38px;
-  width: max-content;
-  transition: background-color 0.2s ease;
-  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.06);
-}
-
-.invalidateButton:hover {
-  background-color: #2563eb;
-}
-
-/* === Button Row: Invalidate + Remove on same line === */
-.buttonRow {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 20px;
-  width: 100%;
-}
-
 /* Floating Remove Icons (×) */
 .removeItem,
 .collectionRemoveId {
@@ -312,7 +274,6 @@
   background-color: #f9fafb;
   border-radius: 8px;
 }
-
 
 /* =======================================
    === Primitive Type Content ===
