+++ frontend/src/features/editors/boxEditor/buttons/ButtonDisplays.module.css
@@ -1,0 +1,46 @@
+.buttonRow {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 20px;
+  width: 100%;
+}
+
+.invalidateButton {
+  background-color: #3b82f6;
+  color: white;
+  padding: 10px 26px;
+  font-size: 0.95rem;
+  border-radius: 8px;
+  border: none;
+  cursor: pointer;
+  height: 38px;
+  width: max-content;
+  transition: background-color 0.2s ease;
+  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.06);
+}
+
+.invalidateButton:hover {
+  background-color: #2563eb;
+}
+
+.removeButton {
+  background-color: #ef4444;
+  color: white;
+}
+
+.removeButton:hover {
+  background-color: #dc2626;
+}
+
+.removeButton {
+  padding: 10px 26px;
+  font-size: 0.95rem;
+  border-radius: 8px;
+  border: none;
+  cursor: pointer;
+  height: 38px;
+  width: max-content;
+  transition: background-color 0.2s ease;
+  box-shadow: 0 1px 2px rgba(0, 0, 0, 0.06);
+}
